refactor (#386)
